requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
caio==0.9.5
click==8.0.4
coverage==6.3.2
dlib==19.23.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.75.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pkg_resources==0.0.0
pluggy==1.0.0
psycopg==3.0.9
psycopg-binary==3.0.9
py==1.11.0
pycodestyle==2.8.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
